+++ virttest/libvirt_xml/devices/librarian.py
@@ -12,7 +12,7 @@ DEVICE_TYPES = ['disk', 'filesystem', 'controller', 'lease',
                 'hostdev', 'redirdev', 'smartcard', 'interface', 'input',
                 'hub', 'graphics', 'video', 'parallel', 'serial', 'console',
                 'channel', 'sound', 'watchdog', 'memballoon', 'rng', 'vsock',
-                'seclabel', 'address', 'emulator', 'panic', 'memory']
+                'seclabel', 'address', 'emulator', 'panic', 'memory', 'filterref']
 
 
 def get(name):
